requirements.txt
pygame-ce
mili-ui
pyserial
# moviepy
# pygamevideo
